requirements.txt
Flask==2.3.3
numpy==1.24.3
pandas==2.0.3
scikit-fuzzy==0.4.2
scikit-learn==1.3.0
matplotlib==3.7.2
python-docx==0.8.11